+++ JsMerge.Core/Enums.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace JsMerge.Core
+﻿namespace JsMerge.Core
 {
 	/// <summary>
 	/// The type of the query result object