+++ JsMerge.Core/Main.cs
@@ -1,8 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using Newtonsoft.Json;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace JsMerge.Core
 {
@@ -14,9 +11,14 @@ namespace JsMerge.Core
 		/// </summary>
 		public static QueryHandler Query { get; private set; }
 		#endregion
+		public static Dictionary<string, MergeConfig>? Config { get; private set; }
 
 		public static string WorkDirectory { get; private set; }
 
+		/// <summary>
+		/// Initializes al required resources
+		/// </summary>
+		/// <param name="workDirectory"></param>
 		public static void Initialize(string workDirectory)
 		{
 			WorkDirectory = workDirectory;
@@ -24,8 +26,50 @@ namespace JsMerge.Core
 			// Create our subsystem objects
 			//
 			Query = new QueryHandler();
+			Config = new Dictionary<string, MergeConfig>();
+
+			// Check if a config file is present
+			//
+			if (File.Exists(WorkDirectory + "/.jsmerge"))
+			{
+				// Load our config
+				using (StreamReader stream = new StreamReader(WorkDirectory + "/.jsmerge"))
+				{
+					Config = JsonConvert.DeserializeObject<Dictionary<string, MergeConfig>>(stream.ReadToEnd());
+				}
+			}
 		}
 
+		/// <summary>
+		/// Creates a new merge file with the given configuration
+		/// </summary>
+		/// <param name="fileName"></param>
+		/// <param name="config"></param>
+		public static MergeResult Merge(string fileName, MergeConfig config)
+		{
+			// Create a new merge result
+			//
+			MergeResult result = new MergeResult(fileName, config);
+
+			// Read all given queries
+			//
+			foreach (string query in config.include)
+			{
+				// Execute our query
+				QueryResult queryResult = ExecuteQuery(query);
+
+				// Add the resulting contents to our merge result
+				result.Append(queryResult);
+			}
+
+			return result;
+		}
+
+		/// <summary>
+		/// Executes a file query string
+		/// </summary>
+		/// <param name="query"></param>
+		/// <returns></returns>
 		public static QueryResult ExecuteQuery(string query)
 		{
 			QueryResult result = QueryResult.Empty;