+++ JsMerge.Core/Util/MergeResult.cs
@@ -1,0 +1,81 @@
+﻿using System.Text;
+
+namespace JsMerge.Core
+{
+	public class MergeResult
+	{
+		private StringBuilder _contents = new StringBuilder();
+		private string _fileName;
+
+		public MergeConfig config;
+		public MergeResult(string fileName, MergeConfig config)
+		{
+			this._fileName = fileName;
+			this.config = config;
+		}
+
+		/// <summary>
+		/// Appends the file contents of the query result to our merge
+		/// </summary>
+		/// <param name="queryResult"></param>
+		public void Append(QueryResult queryResult)
+		{
+			// Preform the correct action for the given result
+			//
+			switch (queryResult.type)
+			{
+				case QueryResultType.FileList:
+					foreach (string file in queryResult.value as string[])
+					{
+						// If in debug mode add the name of the file where the following contents came from
+						//
+						if (config.debug)
+						{
+							_contents.AppendLine("<<<<<<<< " + Path.GetFileName(file));
+						}
+
+						// Read the contents of the file and append it to our merge result
+						//
+						using (StreamReader stream = new StreamReader(file))
+						{
+							_contents.AppendLine(stream.ReadToEnd());
+						}
+					}
+					break;
+
+				case QueryResultType.String:
+					// If in debug mode add the origin where the following contents came from
+					//
+					if (config.debug)
+					{
+						_contents.AppendLine("<<<<<<<< " + queryResult.origin);
+					}
+
+					// If the result contains a string we should add this string to our merge result
+					_contents.AppendLine((string?)queryResult.value);
+					break;
+			}
+		}
+
+		/// <summary>
+		/// Saves our merge result to file
+		/// </summary>
+		public void Save()
+		{
+			// Get out output directory
+			//
+			string dirOut = Main.WorkDirectory;
+			if (config.dirOut != null && config.dirOut != string.Empty)
+			{
+				dirOut += '/' + config.dirOut;
+			}
+
+			// Create a new file with the given name
+			//
+			using (StreamWriter stream = new StreamWriter(dirOut + '/' + _fileName + ".js"))
+			{
+				stream.Write(_contents.ToString());
+			}
+		}
+	}
+}