+++ CHANGES.md
@@ -6,3 +6,8 @@
 - Added: Main class
 - Added: Query system / QueryHandler
 - Added: Getting work directory with path argument(0) if given
+
+### 1.26.06.2022-b
+- Added: .jsmerge config reading
+- Added: Creating merge files
+- Fixed: Not reading files in the correct order